requirements.txt
pydantic
semantic-kernel
azure-identity
microsoft-agents-authentication-msal
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-activity
certifi